Update jam to use store.rs
