+++ src/testutils.rs
@@ -57,3 +57,16 @@ pub mod tmp {
         }
     }
 }
+
+#[cfg(test)]
+pub mod logger {
+    use slog::{o, Drain as _};
+    use slog_term::TestStdoutWriter;
+
+    pub fn test() -> slog::Logger {
+        let decorator = slog_term::PlainSyncDecorator::new(TestStdoutWriter);
+        let drain = slog_term::CompactFormat::new(decorator).build();
+        let drain = std::sync::Mutex::new(drain).fuse();
+        slog::Logger::root(drain, o!())
+    }
+}
